requirements.txt
# LeadGenius Design Suite - Dependencies
# pip install -r requirements.txt

# ─── Core ───
google-genai>=1.0.0
requests>=2.31.0

# ─── Scraping ───
scrapling[all]>=0.4.0

# ─── Email ───
resend>=2.0.0

# ─── Supabase ───
supabase>=2.0.0

# ─── Optional: Screenshot thumbnails ───
# Pillow>=10.0.0
